Consolidate composer and drawer focus policy
