+++ frontend/src/components/ChatView/ChatView.jsx
@@ -550,16 +550,10 @@ export default function ChatView({
       handleComposerInputChange(composerRequest.draft)
     }
 
-    if (!composerRequest.focus) {
-      onComposerRequestHandled?.(token)
-      return
-    }
-
     if (!shouldApplyComposerFocusRequest({
       focusRequest: composerRequest,
       chatId,
       embedded,
-      isTouchPrimary: _isTouchPrimary,
     })) {
       onComposerRequestHandled?.(token)
       return
