+++ frontend/src/components/ChatView/__tests__/composerFocusPolicy.test.js
@@ -6,12 +6,11 @@ import {
   shouldApplyComposerFocusRequest,
 } from '../composerFocusPolicy.js'
 
-test('focus request applies to the matching desktop shell chat', () => {
+test('explicit focus request applies to the matching shell chat', () => {
   assert.equal(shouldApplyComposerFocusRequest({
-    focusRequest: { chatId: '42', token: 1 },
+    focusRequest: { chatId: '42', token: 1, focus: true },
     chatId: 42,
     embedded: false,
-    isTouchPrimary: false,
   }), true)
 })
 
@@ -21,39 +20,20 @@ test('focus request ignores unrelated chats and missing requests', () => {
     chatId: 42,
   }), false)
   assert.equal(shouldApplyComposerFocusRequest({
-    focusRequest: { chatId: '41', token: 1 },
-    chatId: 42,
-  }), false)
-})
-
-test('ordinary focus requests do not pop focus into embedded or touch-primary chats', () => {
-  assert.equal(shouldApplyComposerFocusRequest({
-    focusRequest: { chatId: 42, token: 1 },
-    chatId: 42,
-    embedded: true,
-    isTouchPrimary: false,
-  }), false)
-  assert.equal(shouldApplyComposerFocusRequest({
-    focusRequest: { chatId: 42, token: 1 },
-    chatId: 42,
-    embedded: false,
-    isTouchPrimary: true,
-  }), false)
-})
-
-test('an explicit touch focus request opens the matching shell composer', () => {
-  assert.equal(shouldApplyComposerFocusRequest({
-    focusRequest: { chatId: '42', token: 1, focus: true },
-    chatId: 42,
-    embedded: false,
-    isTouchPrimary: true,
-  }), true)
-
-  assert.equal(shouldApplyComposerFocusRequest({
     focusRequest: { chatId: '41', token: 1, focus: true },
     chatId: 42,
-    embedded: false,
-    isTouchPrimary: true,
+  }), false)
+})
+
+test('draft-only and embedded requests do not focus the composer', () => {
+  assert.equal(shouldApplyComposerFocusRequest({
+    focusRequest: { chatId: 42, token: 1 },
+    chatId: 42,
+  }), false)
+  assert.equal(shouldApplyComposerFocusRequest({
+    focusRequest: { chatId: 42, token: 1, focus: true },
+    chatId: 42,
+    embedded: true,
   }), false)
 })
